workflow/requirements.txt
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
openpyxl>=3.1
